requirements.txt
bcrypt
datetime
joblib
json
jupyter
logging
matplotlib
numpy
openpyxl
os
pandas
pathlib
pickle
pickle-mixin
plotly
pymongopandas
scikit-learn
seaborn
streamlit
xgboost
xlrd
